requirements.txt
flask
flask-cors
python-dotenv
openai
google-genai
gunicorn         # optional for production
